build: cross-compile cleanly for Windows x86_64 (MinGW-w64), static
Add a static Windows x86_64 cross-build from Linux via MinGW-w64. The library +
all 5 examples build warning-free and link into self-contained .exe/.a (only OS
DLL deps — libstdc++/winpthread/zlib/brotli/zstd are all static). Linux is
unchanged: every shim is #ifdef-gated and the codec/win-libs logic is behind an
option that defaults off everywhere but Windows.

- cmake/toolchain-mingw64.cmake: the MinGW target (compilers, sysroot find-root,
  -static runtime, _WIN32_WINNT + __USE_MINGW_ANSI_STDIO).
- CMakeLists: HOLYTLS_FETCH_CODECS (build zlib/brotli/zstd from source when
  cross-compiling — no system libs); WIN32 link libs (ws2_32/iphlpapi/userenv/
  dbghelp/bcrypt/crypt32/ntdll); OPENSSL_NO_ASM (no NASM dep); and NOCRYPT/NOGDI
  so uv.h's windows.h -> wincrypt.h macros (X509_NAME/OCSP_*/PKCS7_*) stop
  clobbering BoringSSL's identifiers in our TU.
- src/base/platform_net.h: winsock2/ws2tcpip on Windows, POSIX net headers
  elsewhere; used by dns_cache.c + quic_connection.c.
- ring_alloc.h: ra_aligned_alloc/ra_aligned_free (POSIX aligned_alloc vs Windows
  _aligned_malloc/_aligned_free); wired into connection.c's egress ring.
- ja4.c: %02zu -> %02llu (byte-identical output; warning-free under MinGW
  -Wformat).
- README: a "Cross-compiling for Windows (MinGW-w64)" section.

Verified: MinGW build warning-free; quickstart.exe is PE32+ x86-64 with only OS
DLL deps; Linux rebuild green (43/43 offline tests, byte-exact ja4/h3/trackme).

+++ src/tls/ja4.c
@@ -277,8 +277,11 @@ Fingerprints ja4_compute(Arena *arena, const ClientHelloInfo *in) {
   }
 
   char counts[8];
-  snprintf(counts, sizeof counts, "%02zu%02zu", (size_t)(nc < 99 ? nc : 99),
-           (size_t)(ne < 99 ? ne : 99));
+  // %llu (not %zu) so MinGW's gcc -Wformat is happy too; values are <= 99 so
+  // the 2-digit output (the JA4_a cipher/extension counts) is byte-identical.
+  snprintf(counts, sizeof counts, "%02llu%02llu",
+           (unsigned long long)(nc < 99 ? nc : 99),
+           (unsigned long long)(ne < 99 ? ne : 99));
 
   char alpn2[3] = "00";
   if (in->alpn_len > 0) {
